fix: move migration script file into database
